Só é possível excluir o estoque de um figurino caso a quantidade de peças entegrue seja igual zero. #396

+++ Codigo/Service/FigurinoService.cs
@@ -148,7 +148,13 @@ namespace Service
         {
             try
             {
-                var estoque = _context.Figurinomanequims.Find(idFigurino, idManequim);
+                var estoque = _context.Figurinomanequims.FindAsync(idFigurino, idManequim).Result;
+
+                if (estoque.QuantidadeEntregue > 0)
+                {
+                    return 400;
+                }
+
                 _context.Figurinomanequims.Remove(estoque);
                 await _context.SaveChangesAsync();
